Use our enhanced `sqlparser-rs` fork
This SQL parser has a dozer specific SQL extension,
which is UDFs with the syntax `function_name<return_type>(arguments)`.

Because the `sqlparser-rs` fork tracks upstream main, and dozer previously
was not using the latest version of `sqlparser-rs`; we need to handle some
(minor) changes that happended in `sqlparser-rs`.

The only change we need to handle so far is the 'pivot' table factor.