fix: Make `Ingestor` async-aware (#2022)
It would have been a small change if not for the snowflake connector.

Turns out the snowflake client we're currently using is not async-aware.
The returned iterator over the table rows from the client is `!Send`.

This forces us to change `Connector::start` to return a `!Send` future,
which in turn results in big refactoring of all the tests.