Reduce memory usage of min and max aggregators (#2123)
This makes use of the fact that the output value of the aggregator will
always be of the same, predetermined type. Therefore, there is no use in
storing the type of the output for every record in the state.

`Field` is 32 bytes, whereas the most common types for the outputs (u64
and float) are 8 bytes. In those cases this changes saves about 3/4 of
the memory.